+++ packages/internal/node/v0.8.0/types.d.ts
@@ -30,7 +30,7 @@ export type CompDefinition = {
         dynamic?: any
     }
     remote?: any
-    hashTag?: '#pre-release' | '#deprecated'
+    hashTag?: '#expreimental' | '#pre-release' | '#deprecated'
     inputs?: NodePort[]
     outputs?: NodePort[]
 }